backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
playwright==1.44.0
gspread==6.1.2
google-auth==2.30.0
pydantic==2.7.1
